refactor: 메뉴와 관련된 클래스들을 menu 패키지로 이동
