+++ src/main/java/kitchenpos/domain/OrderLineItem.java
@@ -9,6 +9,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import kitchenpos.menu.domain.Menu;
 
 @Entity
 public class OrderLineItem {
