+++ src/main/java/kitchenpos/menu/domain/MenuGroup.java
@@ -1,10 +1,11 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import kitchenpos.domain.Name;
 
 @Entity
 public class MenuGroup {
