+++ src/main/java/kitchenpos/menu/domain/MenuProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
