+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -11,17 +11,17 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.application.fixture.MenuFixture;
-import kitchenpos.application.fixture.MenuGroupFixture;
-import kitchenpos.application.fixture.MenuProductFixture;
+import kitchenpos.menu.fixture.MenuFixture;
+import kitchenpos.menu.fixture.MenuGroupFixture;
+import kitchenpos.menu.fixture.MenuProductFixture;
 import kitchenpos.application.fixture.OrderFixture;
 import kitchenpos.application.fixture.OrderLineItemFixture;
 import kitchenpos.application.fixture.ProductFixture;
 import kitchenpos.application.fixture.TableFixture;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.MenuRepository;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderLineItemRepository;
