+++ src/test/java/kitchenpos/application/fixture/OrderLineItemFixture.java
@@ -1,6 +1,6 @@
 package kitchenpos.application.fixture;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.Quantity;
