+++ src/test/java/kitchenpos/menu/fixture/MenuFixture.java
@@ -1,9 +1,9 @@
-package kitchenpos.application.fixture;
+package kitchenpos.menu.fixture;
 
 import java.util.List;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuProduct;
 
 public class MenuFixture {
