+++ src/test/java/kitchenpos/menu/fixture/MenuGroupFixture.java
@@ -1,6 +1,6 @@
-package kitchenpos.application.fixture;
+package kitchenpos.menu.fixture;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 
 public class MenuGroupFixture {
