+++ src/test/java/kitchenpos/menu/fixture/MenuProductFixture.java
@@ -1,7 +1,6 @@
-package kitchenpos.application.fixture;
+package kitchenpos.menu.fixture;
 
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.domain.Product;
 import kitchenpos.domain.Quantity;
